+++ src/components/dashboard/currency.js
@@ -1,6 +1,29 @@
+import { useEffect, useState } from 'react';
 import css from './currency.module.css';
+import { getCurrency } from './currencyAPI/currencyAPI';
+import { nanoid } from 'nanoid';
 
-export const Currency = ({ currencies }) => {
+export const Currency = () => {
+  const [currencies, setCurrencies] = useState(null);
+  useEffect(() => {
+    const getApi = async () => {
+      await getCurrency().then(data => {
+        const currenciesArray = [];
+        data.map(({ ccy, buy, sale }) => {
+          const currenc = {
+            id: nanoid(),
+            ccy,
+            buy,
+            sale,
+          };
+          return currenciesArray.push(currenc);
+        });
+        setCurrencies(currenciesArray);
+        console.log(currenciesArray);
+      });
+    };
+    getApi();
+  }, []);
   return (
     <table className={css.currency}>
       <thead>
@@ -11,13 +34,14 @@ export const Currency = ({ currencies }) => {
         </tr>
       </thead>
       <tbody className={css.currencyBody}>
-        {currencies.map(({ id, ccy, buy, sale }) => (
-          <tr key={id} className={css.currencyInfo}>
-            <td>{ccy}</td>
-            <td>{Number(buy).toFixed(2)}</td>
-            <td>{Number(sale).toFixed(2)}</td>
-          </tr>
-        ))}
+        {currencies &&
+          currencies.map(({ id, ccy, buy, sale }) => (
+            <tr key={id} className={css.currencyInfo}>
+              <td>{ccy}</td>
+              <td>{Number(buy).toFixed(2)}</td>
+              <td>{Number(sale).toFixed(2)}</td>
+            </tr>
+          ))}
       </tbody>
     </table>
   );
